Backend: Filtering based on car brand with/without model done.
Filtering was done using two different ways. Firstly I did it with lambda expression, but with deeper look I saw that there is no need for it, so I switched to utilizing Entity Framework for that job. Never the less I kept part of the code that includes lambda expression.