General: Bugs considering live chat are dealth with
Backend: Added check if user is connected on either of websockets.
If both users are in chat counter for new/unread messages wont update on client side because user is already in chat and is able to see new message. Also those kind of new messages wont come with 'isNew = true' flag. Because user already saw them while chatting with other user.